+++ src/can/mod.rs
@@ -58,6 +58,12 @@ pub trait Error: core::fmt::Debug {
     fn kind(&self) -> ErrorKind;
 }
 
+impl Error for core::convert::Infallible {
+    fn kind(&self) -> ErrorKind {
+        match *self {}
+    }
+}
+
 /// CAN error kind
 ///
 /// This represents a common set of CAN operation errors. HAL implementations are